+++ src/Persistence/src/Dao/SQLite/SQLiteLocationDao.cpp
@@ -53,33 +53,6 @@ SQLiteLocationDao::~SQLiteLocationDao() = default;
 
 bool SQLiteLocationDao::add(Location &location) noexcept
 {
-    // TODO REMOVE ME!!!
-    // TODO We need to activate the "math extensions" for SQLite - the version shipping with Qt does not seem to know about power, sin and cos
-    // https://jonisalonen.com/2014/computing-distance-between-coordinates-can-be-simple-and-fast/
-//    QSqlQuery query1;
-//    query1.setForwardOnly(true);
-//    query1.prepare(
-//        "select power(l.latitude - 48.0, 2) + power((l.longitude - 8.0) * cos(radians( 48.0 )), 2) dist "
-//        "from   location l;"
-//    );
-
-////    query1.bindValue(":latitude", 47.0);
-////    query1.bindValue(":longitude", 8.0);
-//    const bool success = query1.exec();
-//    if (success) {
-
-//        QSqlRecord record = query1.record();
-//        const int distIdx = record.indexOf("dist");
-
-//        while (success && query1.next()) {
-
-//            const double distance = query1.value(distIdx).toDouble();
-//            qDebug() << "SQLiteLocationDao::add: Distance:" << distance;
-//        }
-//    } else {
-//        qDebug() << "SQLiteLocationDao::add: SQL error:" << query1.lastError().text() << "- error code:" << query1.lastError().nativeErrorCode();
-//    }
-
     QSqlQuery query;
     query.prepare(
         "insert into location ("
@@ -192,6 +165,86 @@ bool SQLiteLocationDao::update(const Location &location) noexcept
 #endif
 
     return ok;
+}
+
+std::vector<Location> SQLiteLocationDao::getByPosition(double latitude, double longitude, double distance, bool *ok) const noexcept
+{
+    std::vector<Location> locations;
+    QSqlQuery query;
+    query.setForwardOnly(true);
+
+    // TODO Implement https://jonisalonen.com/2014/computing-distance-between-coordinates-can-be-simple-and-fast/
+    //      SQlite supports "math functions" from Qt 6.5 onwards:
+    //      - https://bugreports.qt.io/browse/QTBUG-108016
+    //      - https://codereview.qt-project.org/c/qt/qtbase/+/440378
+    query.prepare(
+        "select * "
+        "from   location l "
+        "where  l.latitude = :latitude "
+        "  and  l.longitude = :longitude "
+        "order by l.id;"
+    );
+
+    query.bindValue(":latitude", latitude);
+    query.bindValue(":longitude", longitude);
+
+    const bool success = query.exec();
+    if (success) {
+        const bool querySizeFeature = QSqlDatabase::database().driver()->hasFeature(QSqlDriver::QuerySize);
+        if (querySizeFeature) {
+            locations.reserve(query.size());
+        } else {
+            locations.reserve(1);
+        }
+        QSqlRecord record = query.record();
+        const int idIdx = record.indexOf("id");
+        const int titleIdx = record.indexOf("title");
+        const int descriptionIdx = record.indexOf("description");
+        const int typeIdx = record.indexOf("type_id");
+        const int categoryIdx = record.indexOf("category_id");
+        const int countryIdx = record.indexOf("country_id");
+        const int identifierIdx = record.indexOf("identifier");
+        const int latitudeIdx = record.indexOf("latitude");
+        const int longitudeIdx = record.indexOf("longitude");
+        const int altitudeIdx = record.indexOf("altitude");
+        const int pitchIdx = record.indexOf("pitch");
+        const int bankIdx = record.indexOf("bank");
+        const int trueHeadingIdx = record.indexOf("true_heading");
+        const int indicatedAirspeedIdx = record.indexOf("indicated_airspeed");
+        const int onGroundIdx = record.indexOf("on_ground");
+        const int attributesIdx = record.indexOf("attributes");
+
+        while (query.next()) {
+            Location location;
+            location.id = query.value(idIdx).toLongLong();
+            location.title = query.value(titleIdx).toString();
+            location.description = query.value(descriptionIdx).toString();
+            location.typeId = query.value(typeIdx).toLongLong();
+            location.categoryId = query.value(categoryIdx).toLongLong();
+            location.countryId = query.value(countryIdx).toLongLong();
+            location.identifier = query.value(identifierIdx).toString();
+            location.latitude = query.value(latitudeIdx).toDouble();
+            location.longitude = query.value(longitudeIdx).toDouble();
+            location.altitude = query.value(altitudeIdx).toDouble();
+            location.pitch = query.value(pitchIdx).toDouble();
+            location.bank = query.value(bankIdx).toDouble();
+            location.trueHeading = query.value(trueHeadingIdx).toDouble();
+            location.indicatedAirspeed = query.value(indicatedAirspeedIdx).toInt();
+            location.onGround = query.value(onGroundIdx).toBool();
+            location.attributes = query.value(attributesIdx).toLongLong();
+
+            locations.push_back(std::move(location));
+        }
+    }
+#ifdef DEBUG
+    else {
+        qDebug() << "SQLiteLocationDao::getByPosition: SQL error:" << query.lastError().text() << "- error code:" << query.lastError().nativeErrorCode();
+    }
+#endif
+    if (ok != nullptr) {
+        *ok = success;
+    }
+    return locations;
 }
 
 bool SQLiteLocationDao::deleteById(std::int64_t id) noexcept